+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 0.1.0 (2018-10-22)
 
-* Download scrobbled tracks raw data, create an archive on local filesystem for a user - `archive/2`
+* Download scrobbled tracks raw data, create an archive on local filesystem for a user - `archive/1`
 * `extract/5` and `write/2` functions for Lastfm API requests and file output